Last digit

+++ 0x01-variables_if_else_while/1-last_digit.c
@@ -10,7 +10,7 @@
 int main(void)
 {
 	int n;
-	long int lastNum;
+	int lastNum;
 
 	srand(time(0));
 	n = rand() - RAND_MAX / 2;
